+++ R/imports.R
@@ -1,4 +1,4 @@
-#' @importFrom stats na.omit
+#' @importFrom stats na.omit ftable
 NULL
 
 #' @importFrom utils type.convert packageVersion strcapture